# speedup, test and improve datetime parsing