--- a/test-suite/Cargo.toml
+++ b/test-suite/Cargo.toml
@@ -9,6 +9,10 @@ edition = "2018"
 name = "linear"
 harness = false
 
+[[bench]]
+name = "datetime"
+harness = false
+
 [dev-dependencies]
 bencher = "0.1"
 toml = { path = ".." }