@samuelcolvin 2022-05-24
I also noticed that `DatetimeParseError { _private: () }` is quite slow compared to returning an enum member or empty tuple, but I didn't want to change that in case this logic is required?