@epage 2022-09-23
toml-rs has moved to https://github.com/toml-rs/toml_edit.  Feel free to rebase your PR against that repo and open a new issue there.